feat: Extract enums into a new constants file, update `plane.py` to use them.
